refactor: migrate Avatar Badge
