+++ packages/mask/src/plugins/Trader/SNSAdaptor/index.tsx
@@ -15,7 +15,7 @@ import { CrossIsolationMessages, NetworkPluginID, PluginID } from '@masknet/shar
 import type { ChainId } from '@masknet/web3-shared-evm'
 import { SearchResultType } from '@masknet/web3-shared-base'
 import type { Web3Helper } from '@masknet/web3-helpers'
-import { NFTProjectAvatarBadge } from './NFTProjectAvatarBadge.js'
+import { NFTProjectAvatarBadge } from '../../../components/shared/AvatarBadge/NFTProjectAvatarBadge.js'
 import { useCollectionByTwitterHandler } from '../../../plugins/Trader/trending/useTrending.js'
 
 const sns: Plugin.SNSAdaptor.Definition<
