+++ packages/web3-providers/src/helpers/fetchCached.ts
@@ -24,7 +24,7 @@ const RULES = {
     // chainbase ENS
     'https://chainbase-proxy.r2d2.to/v1/ens': Duration.SHORT,
     // Mask Search List
-    'https://raw.githubusercontent.com/DimensionDev/Mask-Search-List': Duration.LONG,
+    'https://dsearch.mask.r2d2.to': Duration.LONG,
 }
 const URLS = Object.keys(RULES) as unknown as Array<keyof typeof RULES>
